+++ finagle-core/src/main/scala/com/twitter/finagle/context/LocalContext.scala
@@ -42,11 +42,13 @@ final class LocalContext private[context] extends Context {
 
   def letClearAll[R](fn: => R): R = local.letClear(fn)
 
-  private[this] def env: Map[Key[_], Any] = local() match {
+  // Exposed for testing
+  private[context] def env: Map[Key[_], Any] = local() match {
     case Some(env) => env
     case None => Map.empty
   }
 
-  private[this] def letLocal[T](env: Map[Key[_], Any])(fn: => T): T =
+  // Exposed for testing
+  private[context] def letLocal[T](env: Map[Key[_], Any])(fn: => T): T =
     local.let(env)(fn)
 }